Fix voucher crud

+++ modules/business/controllers/businesses/vouchers/crud.controller.js
@@ -2,32 +2,31 @@ import fs from "fs";
 import createHttpError from "http-errors";
 
 import prisma from "../../../../../config/prisma.js";
+import { selectParser, includeParser } from "../../../../../helpers/http.helper.js";
 import schemas from "./schemas.js";
+
+const selectWhitelist = [
+    'id', 'campaign_id', 'index', 'type', 'media_url', 'description',
+    'discount_type', 'percent', 'max_value', 'value', 'fixed_price',
+    'usage', 'status', 'collected_count', 'max_use', 'condition_min_bill_value',
+    'condition_beginning_hour', 'condition_ending_hour', 'condition_target'
+];
+
+const includeWhitelist = [
+    'business', 'campaign', 'collectors', 'redeemers', 'appliedProducts'
+]
 
 const getVouchers = async (req, res, next) => {
     try {
         const businessId = req.business.id;
+        const { select, include } = req.query;
+        const selectObject = selectParser(select, selectWhitelist);
+        const includeObject = includeParser(include, includeWhitelist);
         const result = await prisma.voucher.findMany({
             where: { business_id: businessId },
             select: {
-                id: true,
-                campaign_id: true,
-                index: true,
-                type: true,
-                media_url: true,
-                description: true,
-                discount_type: true,
-                percent: true, 
-                max_value: true,
-                value: true,
-                fixed_price: true,
-                usage: true, status: true,
-                collected_count: true,
-                max_use: true,
-                condition_min_bill_value: true,
-                condition_beginning_hour: true,
-                condition_ending_hour: true,
-                condition_target: true,
+                ...selectObject,
+                ...includeObject
             },
         }); res.json(result);
     } catch (e) { next(e); }
@@ -42,6 +41,9 @@ const createVoucher = async (req, res, next) => {
         if (!req.file) {
             throw createHttpError(400);
         }
+        const { select, include } = req.query;
+        const selectObject = selectParser(select, selectWhitelist);
+        const includeObject = includeParser(include, includeWhitelist);
         const businessId = req.business.id;
         const mediaUrl = req.file.path;
         const result = await prisma.voucher.create({
@@ -49,34 +51,19 @@ const createVoucher = async (req, res, next) => {
                 business: {
                     connect: { id: businessId }
                 },
-                media_url: mediaUrl,
+                media_url: `${req.protocol}://${req.get('host')}/${mediaUrl}`,
                 ...body
             },
             select: {
-                id: true,
-                campaign_id: true,
-                index: true,
-                type: true,
-                media_url: true,
-                description: true,
-                discount_type: true,
-                percent: true,
-                max_value: true,
-                value: true,
-                fixed_price: true,
-                usage: true,
-                status: true,
-                collected_count: true,
-                max_use: true,
-                condition_min_bill_value: true,
-                condition_beginning_hour: true,
-                condition_ending_hour: true,
-                condition_target: true,
+                ...selectObject,
+                ...includeObject
             },
         });
         res.json(result);
     } catch (e) {
-        fs.unlinkSync(req.file.path);
+        if (req.file) {
+            fs.unlinkSync(req.file.path);
+        }
         next(e);
     }
 }
@@ -84,29 +71,15 @@ const createVoucher = async (req, res, next) => {
 const getVoucher = async (req, res, next) => {
     try {
         const businessId = req.business.id;
-        const { id } = req.params;
+        const { voucherId } = req.params;
+        const { select, include } = req.query;
+        const selectObject = selectParser(select, selectWhitelist);
+        const includeObject = includeParser(include, includeWhitelist);
         const result = await prisma.voucher.findUnique({
-            where: { id, campaign_id: businessId },
+            where: { id: voucherId, business_id: businessId },
             select: {
-                id: true,
-                campaign_id: true,
-                index: true,
-                type: true,
-                media_url: true,
-                description: true,
-                discount_type: true,
-                percent: true,
-                max_value: true,
-                value: true,
-                fixed_price: true,
-                usage: true,
-                status: true,
-                collected_count: true,
-                max_use: true,
-                condition_min_bill_value: true,
-                condition_beginning_hour: true,
-                condition_ending_hour: true,
-                condition_target: true,
+                ...selectObject,
+                ...includeObject
             },
         });
         res.json(result);
@@ -118,10 +91,13 @@ const getVoucher = async (req, res, next) => {
 const updateVoucher = async (req, res, next) => {
     try {
         const businessId = req.business.id;
-        const { id } = req.params;
+        const { voucherId } = req.params;
+        const { select, include } = req.query;
+        const selectObject = selectParser(select, selectWhitelist);
+        const includeObject = includeParser(include, includeWhitelist);
         const { campaign_id, ...rest } = req.body;
         const result = await prisma.voucher.update({
-            where: { id, campaign_id: businessId },
+            where: { id: voucherId, business_id: businessId },
             data: {
                 campaign: {
                     connect: { id: campaign_id }
@@ -129,25 +105,8 @@ const updateVoucher = async (req, res, next) => {
                 ...rest
             },
             select: {
-                id: true,
-                campaign_id: true,
-                index: true,
-                type: true,
-                media_url: true,
-                description: true,
-                discount_type: true,
-                percent: true,
-                max_value: true,
-                value: true,
-                fixed_price: true,
-                usage: true,
-                status: true,
-                collected_count: true,
-                max_use: true,
-                condition_min_bill_value: true,
-                condition_beginning_hour: true,
-                condition_ending_hour: true,
-                condition_target: true,
+                ...selectObject,
+                ...includeObject
             },
         });
         res.json(result);
@@ -159,33 +118,29 @@ const updateVoucher = async (req, res, next) => {
 const deleteVoucher = async (req, res, next) => {
     try {
         const businessId = req.business.id;
-        const { id } = req.params;
+        const { voucherId } = req.params;
+        const { select, include } = req.query;
+        const selectObject = selectParser(select, selectWhitelist);
+        const includeObject = includeParser(include, includeWhitelist);
+        const voucher = await prisma.voucher.findUnique({
+            where: { id: voucherId, business_id: businessId },
+            select: { media_url: true },
+        })
         const result = await prisma.voucher.delete({
-            where: { id, campaign_id: businessId },
+            where: { id: voucherId, business_id: businessId },
             select: {
-                id: true,
-                campaign_id: true,
-                index: true,
-                type: true,
-                media_url: true,
-                description: true,
-                discount_type: true,
-                percent: true,
-                max_value: true,
-                value: true,
-                fixed_price: true,
-                usage: true,
-                status: true,
-                collected_count: true,
-                max_use: true,
-                condition_min_bill_value: true,
-                condition_beginning_hour: true,
-                condition_ending_hour: true,
-                condition_target: true,
+                ...selectObject,
+                ...includeObject
             },
         });
+        if (voucher.media_url) {
+            const url = new URL(voucher.media_url);
+            const filePath = url.pathname;
+            fs.unlinkSync(`.${filePath}`);
+        }
         res.json(result);
     } catch (e) {
+        console.log(e)
         next(e);
     }
 }
